+++ src/app/factories/pets/useCases/pets.controller.ts
@@ -1,3 +1,4 @@
+import { AppError } from '@domain/errors/AppError';
 import { STATUS } from '@domain/helpers/constants';
 import { formatEmptyValues } from '@domain/helpers/utils/formatEmptyValues';
 import { IPetModel } from '@domain/models/pets';
@@ -18,7 +19,7 @@ class PetsController implements IPetsController {
     const { Location } = await uploadFile(uploadedImage) as unknown as IUploadFile;
     const pet = await PetsService().create({ ...data, image: Location });
 
-    res.status(STATUS.CREATED_CONTENT).json({ pet });
+    res.status(STATUS.CREATED_CONTENT).json(pet);
   }
 
   async findAll(req: Request, res: Response) {
@@ -41,10 +42,15 @@ class PetsController implements IPetsController {
     const { id } = req.params;
 
     const pet = await PetsService().findById(id);
+
+    if (!pet) {
+      throw new AppError('Animal não encontrado', STATUS.NOT_FOUND);
+    }
+
     await PetsService().delete(id);
     splitToDeleteS3Object(pet.image);
 
-    res.status(STATUS.EMPTY_CONTENT).json({ message: 'O animal foi deletado' });
+    res.status(STATUS.EMPTY_CONTENT).send();
   }
 }
